Move docx processing to its own module under Ingestion
