+++ test/text_server/versions_test.exs
@@ -1,7 +1,6 @@
 defmodule TextServer.VersionsTest do
   use TextServer.DataCase
 
-  alias TextServer.TextNodes
   alias TextServer.Versions
   alias TextServer.Versions.Version
 
@@ -186,30 +185,6 @@ defmodule TextServer.VersionsTest do
       assert Enum.count(passage) == 3
       assert List.first(passage).location == [1, 2, 2]
       assert List.last(passage).location == [1, 2, 4]
-    end
-  end
-
-  describe "Version DOCX parsing" do
-    import TextServer.VersionsFixtures
-
-    test "parse_version/1 can parse a docx" do
-      version = version_with_docx_fixture()
-
-      assert {:ok, %Version{} = _} = Versions.parse_version(version)
-
-      text_node = TextNodes.get_by(%{version_id: version.id, location: [1, 1, 1]})
-
-      assert String.contains?(
-               text_node.text,
-               "This is a test document for TextServer.Versions.parse_xml/1"
-             )
-
-      text_node = TextNodes.get_by(%{version_id: version.id, location: [1, 2, 1]})
-
-      assert String.contains?(
-               text_node.text,
-               "Now we want to test inline styles and TextElements."
-             )
     end
   end
 end
